Unify connection setup logic.
